[JIT] Add a python hook for a function to interpret JIT graphs.
This function allows to execute a graph exactly as it is, without going
through a graph executor which would run passes on the graph before
interpreting it. I found this feature extremely helpful when I worked on
a stress-testing script to shake out bugs from the TE fuser: I needed to
execute a very specific set of passes on a graph and nothing else, and
then execute exactly it.

ghstack-source-id: bb266dda707384ddd13288230e76fd9c7c455b3b
Pull Request resolved: https://github.com/pytorch/pytorch/pull/44493